chore(fixed-nav): split demo
