+++ src/packages/__VUE/fixednav/demo/index.vue
@@ -6,6 +6,8 @@
 
     <Overlay />
 
+    <Drag />
+
     <Custom />
   </Demo>
 </template>
@@ -14,5 +16,6 @@
 import Basic from './basic.vue';
 import Left from './left.vue';
 import Overlay from './overlay.vue';
+import Drag from './drag.vue';
 import Custom from './custom.vue';
 </script>
