+++ src/packages/__VUE/fixednav/doc.en-US.md
@@ -16,214 +16,23 @@ app.use(FixedNav);
 
 ### Basic Usage
 
-:::demo
-
-```vue
-<template>
-  <nut-fixed-nav :position="{ top: '70px' }" v-model:visible="visible" :nav-list="navList" />
-</template>
-<script setup>
-import { ref, reactive } from 'vue';
-const visible = ref(false);
-const navList = reactive([
-  {
-    id: 1,
-    text: 'Index',
-    icon: 'https://img11.360buyimg.com/imagetools/jfs/t1/117646/2/11112/1297/5ef83e95E81d77f05/daf8e3b1c81e3c98.png'
-  },
-  {
-    id: 2,
-    text: 'Category',
-    icon: 'https://img12.360buyimg.com/imagetools/jfs/t1/119490/8/9568/1798/5ef83e95E968c69a6/dd029326f7d5042e.png'
-  },
-  {
-    id: 3,
-    text: 'Cart',
-    num: 2,
-    icon: 'https://img14.360buyimg.com/imagetools/jfs/t1/130725/4/3157/1704/5ef83e95Eb976644f/b36c6cfc1cc1a99d.png'
-  },
-  {
-    id: 4,
-    text: 'My',
-    icon: 'https://img12.360buyimg.com/imagetools/jfs/t1/147573/29/1603/1721/5ef83e94E1393a678/5ddf1695ec989373.png'
-  }
-]);
-</script>
-```
-
-:::
+> demo: fixednav basic
 
 ### Left effect
 
-:::demo
+> demo: fixednav left
 
-```vue
-<template>
-  <nut-fixed-nav type="left" :position="{ top: '140px' }" v-model:visible="visible" :nav-list="navList" />
-</template>
-<script setup>
-import { ref, reactive } from 'vue';
-const visible = ref(false);
-const navList = reactive([
-  {
-    id: 1,
-    text: 'Index',
-    icon: 'https://img11.360buyimg.com/imagetools/jfs/t1/117646/2/11112/1297/5ef83e95E81d77f05/daf8e3b1c81e3c98.png'
-  },
-  {
-    id: 2,
-    text: 'Category',
-    icon: 'https://img12.360buyimg.com/imagetools/jfs/t1/119490/8/9568/1798/5ef83e95E968c69a6/dd029326f7d5042e.png'
-  },
-  {
-    id: 3,
-    text: 'Cart',
-    num: 2,
-    icon: 'https://img14.360buyimg.com/imagetools/jfs/t1/130725/4/3157/1704/5ef83e95Eb976644f/b36c6cfc1cc1a99d.png'
-  },
-  {
-    id: 4,
-    text: 'My',
-    icon: 'https://img12.360buyimg.com/imagetools/jfs/t1/147573/29/1603/1721/5ef83e94E1393a678/5ddf1695ec989373.png'
-  }
-]);
-</script>
-```
+### Without Overlay
 
-:::
-
-### Unmask background
-
-:::demo
-
-```vue
-<template>
-  <nut-fixed-nav :overlay="false" :position="{ top: '210px' }" v-model:visible="visible" :nav-list="navList" />
-</template>
-<script setup>
-import { ref, reactive } from 'vue';
-const visible = ref(false);
-const navList = reactive([
-  {
-    id: 1,
-    text: 'Index',
-    icon: 'https://img11.360buyimg.com/imagetools/jfs/t1/117646/2/11112/1297/5ef83e95E81d77f05/daf8e3b1c81e3c98.png'
-  },
-  {
-    id: 2,
-    text: 'Category',
-    icon: 'https://img12.360buyimg.com/imagetools/jfs/t1/119490/8/9568/1798/5ef83e95E968c69a6/dd029326f7d5042e.png'
-  },
-  {
-    id: 3,
-    text: 'Cart',
-    num: 2,
-    icon: 'https://img14.360buyimg.com/imagetools/jfs/t1/130725/4/3157/1704/5ef83e95Eb976644f/b36c6cfc1cc1a99d.png'
-  },
-  {
-    id: 4,
-    text: 'My',
-    icon: 'https://img12.360buyimg.com/imagetools/jfs/t1/147573/29/1603/1721/5ef83e94E1393a678/5ddf1695ec989373.png'
-  }
-]);
-</script>
-```
-
-:::
+> demo: fixednav overlay
 
 ### Support drag and drop with drag
 
-:::demo
-
-```vue
-<template>
-  <nut-drag direction="y" :style="{ right: '0px', bottom: '240px' }">
-    <nut-fixed-nav un-active-text="drag" v-model:visible="visible" :nav-list="navList" />
-  </nut-drag>
-</template>
-<script setup>
-import { ref, reactive } from 'vue';
-const visible = ref(false);
-const navList = reactive([
-  {
-    id: 1,
-    text: 'Index',
-    icon: 'https://img11.360buyimg.com/imagetools/jfs/t1/117646/2/11112/1297/5ef83e95E81d77f05/daf8e3b1c81e3c98.png'
-  },
-  {
-    id: 2,
-    text: 'Category',
-    icon: 'https://img12.360buyimg.com/imagetools/jfs/t1/119490/8/9568/1798/5ef83e95E968c69a6/dd029326f7d5042e.png'
-  },
-  {
-    id: 3,
-    text: 'Cart',
-    num: 2,
-    icon: 'https://img14.360buyimg.com/imagetools/jfs/t1/130725/4/3157/1704/5ef83e95Eb976644f/b36c6cfc1cc1a99d.png'
-  },
-  {
-    id: 4,
-    text: 'My',
-    icon: 'https://img12.360buyimg.com/imagetools/jfs/t1/147573/29/1603/1721/5ef83e94E1393a678/5ddf1695ec989373.png'
-  }
-]);
-</script>
-```
-
-:::
+> demo: fixednav drag
 
 ### Custom use
 
-:::demo
-
-```vue
-<template>
-  <nut-fixed-nav :position="{ top: '280px' }" type="left" v-model:visible="myActive">
-    <template #list>
-      <ul class="nut-fixed-nav__list">
-        <li class="nut-fixed-nav__list-item">1</li>
-        <li class="nut-fixed-nav__list-item">2</li>
-        <li class="nut-fixed-nav__list-item">3</li>
-        <li class="nut-fixed-nav__list-item">4</li>
-        <li class="nut-fixed-nav__list-item">5</li>
-      </ul>
-    </template>
-    <template #btn>
-      <Retweet color="#fff" />
-      <span class="text">{{ myActive ? 'Open' : 'Close' }}</span>
-    </template>
-  </nut-fixed-nav>
-</template>
-<script setup>
-import { ref, reactive } from 'vue';
-const myActive = ref(false);
-const navList = reactive([
-  {
-    id: 1,
-    text: 'Index',
-    icon: 'https://img11.360buyimg.com/imagetools/jfs/t1/117646/2/11112/1297/5ef83e95E81d77f05/daf8e3b1c81e3c98.png'
-  },
-  {
-    id: 2,
-    text: 'Category',
-    icon: 'https://img12.360buyimg.com/imagetools/jfs/t1/119490/8/9568/1798/5ef83e95E968c69a6/dd029326f7d5042e.png'
-  },
-  {
-    id: 3,
-    text: 'Cart',
-    num: 2,
-    icon: 'https://img14.360buyimg.com/imagetools/jfs/t1/130725/4/3157/1704/5ef83e95Eb976644f/b36c6cfc1cc1a99d.png'
-  },
-  {
-    id: 4,
-    text: 'My',
-    icon: 'https://img12.360buyimg.com/imagetools/jfs/t1/147573/29/1603/1721/5ef83e94E1393a678/5ddf1695ec989373.png'
-  }
-]);
-</script>
-```
-
-:::
+> demo: fixednav custom
 
 ## API
