+++ src/packages/__VUE/fixednav/doc.md
@@ -16,215 +16,23 @@ app.use(FixedNav);
 
 ### 基础用法
 
-:::demo
-
-```vue
-<template>
-  <nut-fixed-nav :position="{ top: '70px' }" v-model:visible="visible" :nav-list="navList" />
-</template>
-<script setup>
-import { ref, reactive } from 'vue';
-const visible = ref(false);
-const navList = reactive([
-  {
-    id: 1,
-    text: '首页',
-    icon: 'https://img11.360buyimg.com/imagetools/jfs/t1/117646/2/11112/1297/5ef83e95E81d77f05/daf8e3b1c81e3c98.png'
-  },
-  {
-    id: 2,
-    text: '分类',
-    icon: 'https://img12.360buyimg.com/imagetools/jfs/t1/119490/8/9568/1798/5ef83e95E968c69a6/dd029326f7d5042e.png'
-  },
-  {
-    id: 3,
-    text: '购物车',
-    num: 2,
-    icon: 'https://img14.360buyimg.com/imagetools/jfs/t1/130725/4/3157/1704/5ef83e95Eb976644f/b36c6cfc1cc1a99d.png'
-  },
-  {
-    id: 4,
-    text: '我的',
-    icon: 'https://img12.360buyimg.com/imagetools/jfs/t1/147573/29/1603/1721/5ef83e94E1393a678/5ddf1695ec989373.png'
-  }
-]);
-</script>
-```
-
-:::
+> demo: fixednav basic
 
 ### 左侧效果
 
-:::demo
-
-```vue
-<template>
-  <nut-fixed-nav type="left" :position="{ top: '140px' }" v-model:visible="visible" :nav-list="navList" />
-</template>
-<script setup>
-import { ref, reactive } from 'vue';
-const visible = ref(false);
-const navList = reactive([
-  {
-    id: 1,
-    text: '首页',
-    icon: 'https://img11.360buyimg.com/imagetools/jfs/t1/117646/2/11112/1297/5ef83e95E81d77f05/daf8e3b1c81e3c98.png'
-  },
-  {
-    id: 2,
-    text: '分类',
-    icon: 'https://img12.360buyimg.com/imagetools/jfs/t1/119490/8/9568/1798/5ef83e95E968c69a6/dd029326f7d5042e.png'
-  },
-  {
-    id: 3,
-    text: '购物车',
-    num: 2,
-    icon: 'https://img14.360buyimg.com/imagetools/jfs/t1/130725/4/3157/1704/5ef83e95Eb976644f/b36c6cfc1cc1a99d.png'
-  },
-  {
-    id: 4,
-    text: '我的',
-    icon: 'https://img12.360buyimg.com/imagetools/jfs/t1/147573/29/1603/1721/5ef83e94E1393a678/5ddf1695ec989373.png'
-  }
-]);
-</script>
-```
-
-:::
+> demo: fixednav left
 
 ### 取消背景遮罩
 
-:::demo
-
-```vue
-<template>
-  <nut-fixed-nav :overlay="false" :position="{ top: '210px' }" v-model:visible="visible" :nav-list="navList" />
-</template>
-<script setup>
-import { ref, reactive } from 'vue';
-const visible = ref(false);
-const navList = reactive([
-  {
-    id: 1,
-    text: '首页',
-    icon: 'https://img11.360buyimg.com/imagetools/jfs/t1/117646/2/11112/1297/5ef83e95E81d77f05/daf8e3b1c81e3c98.png'
-  },
-  {
-    id: 2,
-    text: '分类',
-    icon: 'https://img12.360buyimg.com/imagetools/jfs/t1/119490/8/9568/1798/5ef83e95E968c69a6/dd029326f7d5042e.png'
-  },
-  {
-    id: 3,
-    text: '购物车',
-    num: 2,
-    icon: 'https://img14.360buyimg.com/imagetools/jfs/t1/130725/4/3157/1704/5ef83e95Eb976644f/b36c6cfc1cc1a99d.png'
-  },
-  {
-    id: 4,
-    text: '我的',
-    icon: 'https://img12.360buyimg.com/imagetools/jfs/t1/147573/29/1603/1721/5ef83e94E1393a678/5ddf1695ec989373.png'
-  }
-]);
-</script>
-```
-
-:::
+> demo: fixednav overlay
 
 ### 配合 Drag 支持拖拽
 
-:::demo
-
-```vue
-<template>
-  <nut-drag direction="y" :style="{ right: '0px', bottom: '240px' }">
-    <nut-fixed-nav un-active-text="支持拖拽" v-model:visible="visible" :nav-list="navList" />
-  </nut-drag>
-</template>
-<script setup>
-import { ref, reactive } from 'vue';
-const visible = ref(false);
-const navList = reactive([
-  {
-    id: 1,
-    text: '首页',
-    icon: 'https://img11.360buyimg.com/imagetools/jfs/t1/117646/2/11112/1297/5ef83e95E81d77f05/daf8e3b1c81e3c98.png'
-  },
-  {
-    id: 2,
-    text: '分类',
-    icon: 'https://img12.360buyimg.com/imagetools/jfs/t1/119490/8/9568/1798/5ef83e95E968c69a6/dd029326f7d5042e.png'
-  },
-  {
-    id: 3,
-    text: '购物车',
-    num: 2,
-    icon: 'https://img14.360buyimg.com/imagetools/jfs/t1/130725/4/3157/1704/5ef83e95Eb976644f/b36c6cfc1cc1a99d.png'
-  },
-  {
-    id: 4,
-    text: '我的',
-    icon: 'https://img12.360buyimg.com/imagetools/jfs/t1/147573/29/1603/1721/5ef83e94E1393a678/5ddf1695ec989373.png'
-  }
-]);
-</script>
-```
-
-:::
+> demo: fixednav drag
 
 ### 自定义使用
 
-:::demo
-
-```vue
-<template>
-  <nut-fixed-nav :position="{ top: '280px' }" type="left" v-model:visible="myActive">
-    <template #list>
-      <ul class="nut-fixed-nav__list">
-        <li class="nut-fixed-nav__list-item">1</li>
-        <li class="nut-fixed-nav__list-item">2</li>
-        <li class="nut-fixed-nav__list-item">3</li>
-        <li class="nut-fixed-nav__list-item">4</li>
-        <li class="nut-fixed-nav__list-item">5</li>
-      </ul>
-    </template>
-    <template #btn>
-      <Retweet color="#fff" />
-      <span class="text">{{ myActive ? '自定义开' : '自定义关' }}</span>
-    </template>
-  </nut-fixed-nav>
-</template>
-<script setup>
-import { ref, reactive } from 'vue';
-import { Retweet } from '@nutui/icons-vue';
-const myActive = ref(false);
-const navList = reactive([
-  {
-    id: 1,
-    text: '首页',
-    icon: 'https://img11.360buyimg.com/imagetools/jfs/t1/117646/2/11112/1297/5ef83e95E81d77f05/daf8e3b1c81e3c98.png'
-  },
-  {
-    id: 2,
-    text: '分类',
-    icon: 'https://img12.360buyimg.com/imagetools/jfs/t1/119490/8/9568/1798/5ef83e95E968c69a6/dd029326f7d5042e.png'
-  },
-  {
-    id: 3,
-    text: '购物车',
-    num: 2,
-    icon: 'https://img14.360buyimg.com/imagetools/jfs/t1/130725/4/3157/1704/5ef83e95Eb976644f/b36c6cfc1cc1a99d.png'
-  },
-  {
-    id: 4,
-    text: '我的',
-    icon: 'https://img12.360buyimg.com/imagetools/jfs/t1/147573/29/1603/1721/5ef83e94E1393a678/5ddf1695ec989373.png'
-  }
-]);
-</script>
-```
-
-:::
+> demo: fixednav custom
 
 ## API
